adds nyc, splits test file, updates mocha, DEP0005
