+++ docs/NoteOnEvent.html
@@ -99,7 +99,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="note-events_note-on-event.js.html">note-events/note-on-event.js</a>, <a href="note-events_note-on-event.js.html#line8">line 8</a>
+        <a href="note-events_note-on-event.js.html">note-events/note-on-event.js</a>, <a href="note-events_note-on-event.js.html#line9">line 9</a>
     </li></ul></dd>
     
 
@@ -504,7 +504,7 @@ Note on at channel 0 is 0x90 (144)
 <br class="clear">
 
 <footer>
-    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Fri Mar 15 2019 08:23:19 GMT-0700 (PDT) using the Minami theme.
+    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.6.3</a> on Thu Jul 18 2019 23:31:03 GMT+0100 (British Summer Time) using the Minami theme.
 </footer>
 
 <script>prettyPrint();</script>
