+++ docs/ProgramChangeEvent.html
@@ -99,7 +99,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="meta-events_program-change-event.js.html">meta-events/program-change-event.js</a>, <a href="meta-events_program-change-event.js.html#line9">line 9</a>
+        <a href="meta-events_program-change-event.js.html">meta-events/program-change-event.js</a>, <a href="meta-events_program-change-event.js.html#line10">line 10</a>
     </li></ul></dd>
     
 
@@ -238,7 +238,7 @@
 <br class="clear">
 
 <footer>
-    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Fri Mar 15 2019 08:23:19 GMT-0700 (PDT) using the Minami theme.
+    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.6.3</a> on Thu Jul 18 2019 23:31:03 GMT+0100 (British Summer Time) using the Minami theme.
 </footer>
 
 <script>prettyPrint();</script>
