+++ docs/Track.html
@@ -99,7 +99,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="track.js.html">track.js</a>, <a href="track.js.html#line21">line 21</a>
+        <a href="track.js.html">track.js</a>, <a href="track.js.html#line22">line 22</a>
     </li></ul></dd>
     
 
@@ -2772,7 +2772,7 @@ Events with a specific startTick property are set aside for now will be merged i
 <br class="clear">
 
 <footer>
-    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Fri Mar 15 2019 08:23:19 GMT-0700 (PDT) using the Minami theme.
+    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.6.3</a> on Thu Jul 18 2019 23:31:03 GMT+0100 (British Summer Time) using the Minami theme.
 </footer>
 
 <script>prettyPrint();</script>
