+++ docs/Writer.html
@@ -276,7 +276,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="writer.js.html">writer.js</a>, <a href="writer.js.html#line41">line 41</a>
+        <a href="writer.js.html">writer.js</a>, <a href="writer.js.html#line40">line 40</a>
     </li></ul></dd>
     
 
@@ -380,7 +380,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="writer.js.html">writer.js</a>, <a href="writer.js.html#line28">line 28</a>
+        <a href="writer.js.html">writer.js</a>, <a href="writer.js.html#line27">line 27</a>
     </li></ul></dd>
     
 
@@ -484,7 +484,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="writer.js.html">writer.js</a>, <a href="writer.js.html#line50">line 50</a>
+        <a href="writer.js.html">writer.js</a>, <a href="writer.js.html#line49">line 49</a>
     </li></ul></dd>
     
 
@@ -588,7 +588,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="writer.js.html">writer.js</a>, <a href="writer.js.html#line66">line 66</a>
+        <a href="writer.js.html">writer.js</a>, <a href="writer.js.html#line65">line 65</a>
     </li></ul></dd>
     
 
@@ -723,7 +723,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="writer.js.html">writer.js</a>, <a href="writer.js.html#line58">line 58</a>
+        <a href="writer.js.html">writer.js</a>, <a href="writer.js.html#line57">line 57</a>
     </li></ul></dd>
     
 
@@ -796,7 +796,7 @@
 <br class="clear">
 
 <footer>
-    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Fri Mar 15 2019 08:23:19 GMT-0700 (PDT) using the Minami theme.
+    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.6.3</a> on Thu Jul 18 2019 23:31:03 GMT+0100 (British Summer Time) using the Minami theme.
 </footer>
 
 <script>prettyPrint();</script>
