+++ docs/controller-change-event.js.html
@@ -40,6 +40,7 @@
     <section>
         <article>
             <pre class="prettyprint source linenums"><code>import {Constants} from './constants';
+import {Utils} from './utils.js';
 
 /**
  * Holds all data for a "controller change" MIDI event
@@ -67,7 +68,7 @@ export {ControllerChangeEvent};
 <br class="clear">
 
 <footer>
-    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Fri Mar 15 2019 08:23:19 GMT-0700 (PDT) using the Minami theme.
+    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.6.3</a> on Thu Jul 18 2019 23:31:03 GMT+0100 (British Summer Time) using the Minami theme.
 </footer>
 
 <script>prettyPrint();</script>
