+++ docs/note-events_note-event.js.html
@@ -54,11 +54,11 @@ class NoteEvent {
 		// Set default fields
 		fields = Object.assign({
 			channel: 1,
-		    repeat: 1,
-		    sequential: false,
-		    startTick: null,
-		    velocity: 50,
-		    wait: 0,
+			repeat: 1,
+			sequential: false,
+			startTick: null,
+			velocity: 50,
+			wait: 0,
 		}, fields);
 
 		this.data 		= [];
@@ -203,7 +203,7 @@ class NoteEvent {
 		}
 
 		return this;
-	};
+	}
 }
 
 export {NoteEvent};
@@ -219,7 +219,7 @@ export {NoteEvent};
 <br class="clear">
 
 <footer>
-    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Fri Mar 15 2019 08:23:19 GMT-0700 (PDT) using the Minami theme.
+    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.6.3</a> on Thu Jul 18 2019 23:31:03 GMT+0100 (British Summer Time) using the Minami theme.
 </footer>
 
 <script>prettyPrint();</script>
