+++ docs/note-events_note-off-event.js.html
@@ -52,7 +52,7 @@ class NoteOffEvent {
 		fields = Object.assign({
 			channel: 1,
 			noteOnTick: null,
-		    velocity: 50,
+			velocity: 50,
 		}, fields);
 
 		this.type 		= 'note-off';
@@ -65,7 +65,7 @@ class NoteOffEvent {
 		this.midiNumber = Utils.getPitch(this.pitch);
 		this.tick 		= null;
 		this.delta 		= Utils.getTickDuration(this.duration);
-		this.data 		= fields.data;	
+		this.data 		= fields.data;
 	}
 
 	/**
@@ -113,7 +113,7 @@ export {NoteOffEvent};
 <br class="clear">
 
 <footer>
-    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Fri Mar 15 2019 08:23:19 GMT-0700 (PDT) using the Minami theme.
+    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.6.3</a> on Thu Jul 18 2019 23:31:03 GMT+0100 (British Summer Time) using the Minami theme.
 </footer>
 
 <script>prettyPrint();</script>
