+++ docs/note-events_note-on-event.js.html
@@ -52,8 +52,8 @@ class NoteOnEvent {
 		fields = Object.assign({
 			channel: 1,
 			startTick: null,
-		    velocity: 50,
-		    wait: 0,
+			velocity: 50,
+			wait: 0,
 		}, fields);
 
 		this.type 		= 'note-on';
@@ -123,7 +123,7 @@ export {NoteOnEvent};
 <br class="clear">
 
 <footer>
-    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Fri Mar 15 2019 08:23:19 GMT-0700 (PDT) using the Minami theme.
+    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.6.3</a> on Thu Jul 18 2019 23:31:03 GMT+0100 (British Summer Time) using the Minami theme.
 </footer>
 
 <script>prettyPrint();</script>
