+++ docs/utils.js.html
@@ -74,14 +74,14 @@ class Utils {
 	}
 
 	/**
-     * Returns the correct MIDI number for the specified pitch.
-     * Uses Tonal Midi - https://github.com/danigb/tonal/tree/master/packages/midi
-     * @param {(string|number)} pitch - 'C#4' or midi note code
-     * @return {number}
-     */
-     static getPitch(pitch) {
-     	return toMidi(pitch);
-     }
+	 * Returns the correct MIDI number for the specified pitch.
+	 * Uses Tonal Midi - https://github.com/danigb/tonal/tree/master/packages/midi
+	 * @param {(string|number)} pitch - 'C#4' or midi note code
+	 * @return {number}
+	 */
+	static getPitch(pitch) {
+		return toMidi(pitch);
+	}
 
 	/**
 	 * Translates number of ticks to MIDI timestamp format, returning an array of
@@ -93,22 +93,22 @@ class Utils {
 	 * @return {array} - Bytes that form the MIDI time value
 	 */
 	static numberToVariableLength(ticks) {
-	    var buffer = ticks &amp; 0x7F;
+		var buffer = ticks &amp; 0x7F;
 
-	    while (ticks = ticks >> 7) {
-	        buffer &lt;&lt;= 8;
-	        buffer |= ((ticks &amp; 0x7F) | 0x80);
-	    }
+		while (ticks = ticks >> 7) {
+			buffer &lt;&lt;= 8;
+			buffer |= ((ticks &amp; 0x7F) | 0x80);
+		}
 
-	    var bList = [];
-	    while (true) {
-	        bList.push(buffer &amp; 0xff);
+		var bList = [];
+		while (true) {
+			bList.push(buffer &amp; 0xff);
 
-	        if (buffer &amp; 0x80) buffer >>= 8
-	        else { break; }
-	    }
+			if (buffer &amp; 0x80) buffer >>= 8
+			else { break; }
+		}
 
-	    return bList;
+		return bList;
 	}
 
 	/**
@@ -172,7 +172,7 @@ class Utils {
 		return hexArray;
 	}
 
-	/**	
+	/**
 	 * Converts value to array if needed.
 	 * @param {string} value
 	 * @return {array}
@@ -191,7 +191,7 @@ class Utils {
 		// Max passed value limited to 100
 		velocity = velocity > 100 ? 100 : velocity;
 		return Math.round(velocity / 100 * 127);
-	};
+	}
 
 	/**
 	 * Gets the total number of ticks of a specified duration.
@@ -237,21 +237,27 @@ class Utils {
 				return 4;
 			case '2':
 				return 2;
-			case 'd2':
+			case 'd2': // Dotted half
 				return 3;
+			case 'dd2': // Double dotted half
+				return 3.5;
 			case '4':
 				return 1;
 			case '4t':
 				return 0.666;
-			case 'd4':
+			case 'd4': // Dotted quarter
 				return 1.5;
+			case 'dd4': // Double dotted quarter
+				return 1.75;
 			case '8':
 				return 0.5;
 			case '8t':
 				// For 8th triplets, let's divide a quarter by 3, round to the nearest int, and substract the remainder to the last one.
 				return 0.33;
-			case 'd8':
+			case 'd8': // Dotted eighth
 				return 0.75;
+			case 'dd8': // Double dotted eighth
+				return 0.875;
 			case '16':
 				return 0.25;
 			case '16t':
@@ -266,7 +272,7 @@ class Utils {
 		}
 
 		throw duration + ' is not a valid duration.';
-	};
+	}
 }
 
 export {Utils};
@@ -282,7 +288,7 @@ export {Utils};
 <br class="clear">
 
 <footer>
-    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Fri Mar 15 2019 08:23:19 GMT-0700 (PDT) using the Minami theme.
+    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.6.3</a> on Thu Jul 18 2019 23:31:03 GMT+0100 (British Summer Time) using the Minami theme.
 </footer>
 
 <script>prettyPrint();</script>
