+++ docs/vexflow.js.html
@@ -43,7 +43,7 @@
 import {Track} from './track';
 
 class VexFlow {
-	
+
 	constructor() {
 		// code...
 	}
@@ -73,7 +73,7 @@ class VexFlow {
 			}
 
 			track.addEvent(new NoteEvent({pitch: pitches, duration: this.convertDuration(tickable), wait: wait}));
-			
+
 			// reset wait
 			wait = 0;
 		});
@@ -88,7 +88,7 @@ class VexFlow {
 	 */
 	convertPitch(pitch) {
 		return pitch.replace('/', '');
-	} 
+	}
 
 
 	/**
@@ -108,7 +108,7 @@ class VexFlow {
 		}
 
 		return note.duration;
-	};
+	}
 }
 
 export {VexFlow};
@@ -124,7 +124,7 @@ export {VexFlow};
 <br class="clear">
 
 <footer>
-    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Fri Mar 15 2019 08:23:19 GMT-0700 (PDT) using the Minami theme.
+    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.6.3</a> on Thu Jul 18 2019 23:31:03 GMT+0100 (British Summer Time) using the Minami theme.
 </footer>
 
 <script>prettyPrint();</script>
