+++ docs/writer.js.html
@@ -40,7 +40,6 @@
     <section>
         <article>
             <pre class="prettyprint source linenums"><code>import {HeaderChunk} from './header-chunk';
-import {Constants} from './constants';
 import {Utils} from './utils';
 
 /**
@@ -53,7 +52,7 @@ class Writer {
 		// Ensure track is an array
 		tracks = Utils.toArray(tracks);
 
-		this.data = [];		
+		this.data = [];
 		this.data.push(new HeaderChunk(tracks.length))
 
 		// For each track add final end of track event and build data
@@ -89,7 +88,7 @@ class Writer {
      * @return {string}
      */
     dataUri() {
-    	return 'data:audio/midi;base64,' + this.base64();
+		return 'data:audio/midi;base64,' + this.base64();
     }
 
 	/**
@@ -97,7 +96,7 @@ class Writer {
 	 * @return {string}
 	 */
     stdout() {
-    	return process.stdout.write(new Buffer(this.buildFile()));
+		return process.stdout.write(new Buffer(this.buildFile()));
     }
 
 	/**
@@ -106,7 +105,7 @@ class Writer {
 	 */
 	saveMIDI(filename) {
 		const fs = require('fs');
-		const buffer = new Buffer(this.buildFile());
+		const buffer = new Buffer.from(this.buildFile());
 		fs.writeFile(filename + '.mid', buffer, function (err) {
 			if(err) throw err;
 		});
@@ -126,7 +125,7 @@ export {Writer};
 <br class="clear">
 
 <footer>
-    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Fri Mar 15 2019 08:23:19 GMT-0700 (PDT) using the Minami theme.
+    Generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.6.3</a> on Thu Jul 18 2019 23:31:03 GMT+0100 (British Summer Time) using the Minami theme.
 </footer>
 
 <script>prettyPrint();</script>
